requirements.txt
google.generativeai
moviepy
gTTs
